Refactor to use shoes-core specs
Now the shoes-core specs are copied directly. We can maintain a copy
here for the time being, with modifications as necessary to run with
these backends.

Specs are now opt-in by file (see spec/browser/browser_runner.rb). To
exclude a spec from running in the browser suite, mark it :no_browser.