Improve build/bootstrap
While updating the build, I attempted to update the Shoes 4 gem to the
latest commit on master, but recent changes to initialization don't work
properly with Opal. Spent hours trying to find the root cause, with no
luck. Seems that a Flow (the top_slot) is getting instantiated without
calling Shoes::Common::Initialization#initialize. No idea why. All of my
attempts to come up with a simpler failing case failed to fail.

One interesting thing is that in MRI, a class that includes a module
that includes another module has the "another module" in its list of
ancestors, whereas in Opal it does not. Still, Opal does seem to support
including an #initialize method defined in another module.

Last good commit on Shoes 4 for the time being is 6f2d017.